# Replace module name check linear scan with map lookup