--- a/internal/wasm/namespace_test.go
+++ b/internal/wasm/namespace_test.go
@@ -23,7 +23,8 @@ func TestNamespace_addModule(t *testing.T) {
 
 		require.Equal(t, map[string]*ModuleInstance{m1.Name: m1}, ns.modules)
 		// Doesn't affect module names
-		require.Nil(t, ns.moduleNames)
+		require.Zero(t, len(ns.moduleNamesSet))
+		require.Nil(t, ns.moduleNamesList)
 	})
 
 	t.Run("redundant ok", func(t *testing.T) {
@@ -46,7 +47,8 @@ func TestNamespace_deleteModule(t *testing.T) {
 
 		// Leaves the other module alone
 		require.Equal(t, map[string]*ModuleInstance{m1.Name: m1}, ns.modules)
-		require.Equal(t, []string{m1.Name}, ns.moduleNames)
+		require.Equal(t, map[string]struct{}{m1.Name: {}}, ns.moduleNamesSet)
+		require.Equal(t, []string{m1.Name}, ns.moduleNamesList)
 	})
 
 	t.Run("ok if missing", func(t *testing.T) {
@@ -57,7 +59,8 @@ func TestNamespace_deleteModule(t *testing.T) {
 		ns.deleteModule(m1.Name)
 
 		require.Zero(t, len(ns.modules))
-		require.Zero(t, len(ns.moduleNames))
+		require.Zero(t, len(ns.moduleNamesSet))
+		require.Zero(t, len(ns.moduleNamesList))
 	})
 }
 
@@ -90,14 +93,15 @@ func TestNamespace_requireModules(t *testing.T) {
 }
 
 func TestNamespace_requireModuleName(t *testing.T) {
-	ns := &Namespace{}
+	ns := &Namespace{moduleNamesSet: map[string]struct{}{}}
 
 	t.Run("first", func(t *testing.T) {
 		err := ns.requireModuleName("m1")
 		require.NoError(t, err)
 
 		// Ensure it adds the module name, and doesn't impact the module list.
-		require.Equal(t, []string{"m1"}, ns.moduleNames)
+		require.Equal(t, []string{"m1"}, ns.moduleNamesList)
+		require.Equal(t, map[string]struct{}{"m1": {}}, ns.moduleNamesSet)
 		require.Zero(t, len(ns.modules))
 	})
 	t.Run("second", func(t *testing.T) {
@@ -105,7 +109,8 @@ func TestNamespace_requireModuleName(t *testing.T) {
 		require.NoError(t, err)
 
 		// Appends in order.
-		require.Equal(t, []string{"m1", "m2"}, ns.moduleNames)
+		require.Equal(t, []string{"m1", "m2"}, ns.moduleNamesList)
+		require.Equal(t, map[string]struct{}{"m1": {}, "m2": {}}, ns.moduleNamesSet)
 	})
 	t.Run("existing", func(t *testing.T) {
 		err := ns.requireModuleName("m2")
@@ -121,7 +126,8 @@ func TestNamespace_AliasModule(t *testing.T) {
 	ns.AliasModule("m1", "m2")
 	require.Equal(t, map[string]*ModuleInstance{"m1": m1, "m2": m1}, ns.modules)
 	// Doesn't affect module names
-	require.Nil(t, ns.moduleNames)
+	require.Zero(t, len(ns.moduleNamesSet))
+	require.Nil(t, ns.moduleNamesList)
 }
 
 func TestNamespace_CloseWithExitCode(t *testing.T) {
@@ -158,7 +164,8 @@ func TestNamespace_CloseWithExitCode(t *testing.T) {
 
 			// Namespace state zeroed
 			require.Zero(t, len(ns.modules))
-			require.Zero(t, len(ns.moduleNames))
+			require.Zero(t, len(ns.moduleNamesSet))
+			require.Zero(t, len(ns.moduleNamesList))
 		})
 	}
 
@@ -183,7 +190,8 @@ func TestNamespace_CloseWithExitCode(t *testing.T) {
 
 		// Namespace state zeroed
 		require.Zero(t, len(ns.modules))
-		require.Zero(t, len(ns.moduleNames))
+		require.Zero(t, len(ns.moduleNamesSet))
+		require.Zero(t, len(ns.moduleNamesList))
 	})
 }
 
@@ -209,6 +217,7 @@ func newTestNamespace() (*Namespace, *ModuleInstance, *ModuleInstance) {
 	m2.CallCtx = NewCallContext(ns, m2, nil)
 
 	ns.modules = map[string]*ModuleInstance{m1.Name: m1, m2.Name: m2}
-	ns.moduleNames = []string{m1.Name, m2.Name}
+	ns.moduleNamesSet = map[string]struct{}{m1.Name: {}, m2.Name: {}}
+	ns.moduleNamesList = []string{m1.Name, m2.Name}
 	return ns, m1, m2
 }